Fix commandOptionsParser and move to bot package
