+++ packages/bot/src/index.ts
@@ -3,6 +3,7 @@ export * from '@discordeno/rest'
 export * from '@discordeno/types'
 export * from '@discordeno/utils'
 export * from './bot.js'
+export * from './commandOptionsParser.js'
 export * from './constants.js'
 export * from './handlers/index.js'
 export * from './handlers.js'
