+++ packages/utils/src/index.ts
@@ -5,7 +5,6 @@ export * from './casing.js'
 export * from './colors.js'
 export * from './hash.js'
 export * from './images.js'
-export * from './interactions.js'
 export * from './logger.js'
 export * from './oauth2.js'
 export * from './permissions.js'
